Clean up clojure-mode faces
This removes most of the faces (as they don't appear to be used by
clojure-mode any more) and sets clojure-keyword-face to be base0E (as
recommended by the styling guidelines).

We don't currently set clojure-character-face or
clojure-interop-method-face because I'm not sure what the best colors
would be for those and they already inherit from themed faces.

Closes #40