Unification of colour and settings application.
This change combines the existing colour key transformation with the
settings application. The hope is that this makes it easier to add new
user settings in the future. I had wanted to extend the mode line
emphasis to allow a choice of either a box or a darker foreground
colour. This would have meant putting the setting specific code in two
separate functions (settings for the box and colour-key for the
foreground). Before I add that option I figured this refactoring would
be worthwhile to try out. I've tested it with the default and tomorrow
themes with both custom settings and the behaviour is identical to
what is currently on master.