Use base02 for `magit-diff-hunk-heading-highlight`
The `-highlight` face is used for the currently selected hunk in the
magit diff and revision views. It should be distinct from the non `-
highlight` font to aid in naviation.

Per [the guidelines](https://github.com/chriskempson/base16/blob/main/
styling.md), base02 is to be used for "Selection background," which
seems apropos to me. It also looks much better in my emacs setup using
base16-framer, for whatever that's worth.